% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BigQuery_setup.R
\name{bigquery_setup_server}
\alias{bigquery_setup_server}
\title{BigQuery Setup Server}
\usage{
bigquery_setup_server(
  id,
  secrets_json = "~/.shinyBigQuery/client_secret/client_secret.json"
)
}
\arguments{
\item{id}{The Module namespace}

\item{secrets_json}{Location of Google secrets json}
}
\value{
BigQuery connection variables and user information
}
\description{
BigQuery Setup Server
}
