% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{A series of options/params to be used inside the app including:
\itemize{
\item{secrets_json: A string, containing a file path to a Google OAuth 2.0 Client}
}}
}
\value{
No return value, called to start the Shiny Application!
}
\description{
Run the Shiny Application
}
