% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{package}
\name{shinyBigQuery}
\alias{shinyBigQuery}
\title{shinyBigQuery: Access Google BigQuery through R Shiny}
\description{
A shiny module to authenticate your R Shiny Application with Google BigQuery.
}
